Don't replace the whole `ENV` unconditionally
All we want to do is to add some keys with backup of the original values
of bundler-related variables.

Apparently fully replacing the ENV break causes issues when loading
shared libraries where it is assumed that ENV should inmediately follow
ARGV in memory. For example, the go runtime.

Not really sure this is our issue, but replacing the whole `ENV` is
unnecessary anyways, so we can simplify that and get the issue fixed as
a result.