Lazily load `optparse`
The `optparse` library is loaded by `rubygems/command`, which should
only be loaded when using the `gem` CLI. However, it was also being
loaded in some other cases.

This commit tries to lazily load it, so that it doesn't interfere with
other usages. For example, when `rubygems/installer` is required
directly from `bundler`.